[ucd/name] Make library no_std
